Moves validateFolderName to messaging/utils/validations.js.
Closes department-of-veterans-affairs/messaging-fe#99.
