+++ src/js/messaging/components/ModalCreateFolder.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import ErrorableTextInput from '../../common/components/form-elements/ErrorableTextInput';
 import Modal from '../../common/components/Modal';
 import { makeField } from '../../common/model/fields';
-import { validateIfDirty, isNotBlank } from '../../common/utils/validations';
+import { validateFolderName } from '../utils/validations';
 import { createNewFolderSettings } from '../config';
 
 class ModalCreateFolder extends React.Component {
@@ -11,7 +11,6 @@ class ModalCreateFolder extends React.Component {
     super(props);
     this.handleSubmit = this.handleSubmit.bind(this);
     this.handleValueChange = this.handleValueChange.bind(this);
-    this.validateFolderName = this.validateFolderName.bind(this);
   }
 
   handleValueChange(field) {
@@ -31,40 +30,10 @@ class ModalCreateFolder extends React.Component {
     }
   }
 
-  validateFolderName(folderName, existingFolders = []) {
-    const err = {};
-    // TODO: Refactor isNotBlank validator to trim input.
-    const trimmedFolderName = makeField(folderName.value.trim(), folderName.dirty);
-
-    if (!validateIfDirty(trimmedFolderName, isNotBlank)) {
-      err.hasError = true;
-      err.type = 'empty';
-    }
-
-    // Disallows anything other than a-z, 0-9, and space
-    // (case insensitive)
-    const allowedRegExp = /[^a-z0-9\s]/ig;
-    if (allowedRegExp.test(trimmedFolderName.value)) {
-      err.hasError = true;
-      err.type = 'patternMismatch';
-    }
-
-    const doesFolderExist = (folder) => {
-      return trimmedFolderName.value.toLowerCase() === folder.name.toLowerCase();
-    };
-
-    if (!!existingFolders.find(doesFolderExist)) {
-      err.hasError = true;
-      err.type = 'exists';
-    }
-
-    return err;
-  }
-
   render() {
     const foldersWeHave = this.props.folders;
     const newFolderName = this.props.newFolderName;
-    const error = this.validateFolderName(newFolderName, foldersWeHave);
+    const error = validateFolderName(newFolderName, foldersWeHave);
 
     const modalContents = (
       <form onSubmit={this.handleSubmit}>
